[ALPS02767202] DMA: get the sgs length from the macro
[Detail]
pseudo m4u store the s_length in sg_dma_len, it may be in different
field depend on the CONFIG_NEED_SG_DMA_LENGTH, get the length from
the sg_dma_len macro.

For normal path, it will go through the if condition which is the same
with the origin routine. Only for pseudo m4u users which allocate mva
with the memory has PFNMAP flags set and has no page struct will go
through the else condition, its OK for the normal path.

[Solution]
Get the sgs length from the sg_dma_len macro if we store the length
in there.

MTK-Commit-Id: 8f6575bd7f65b86b5034f186f9b50c4479f15e7b

Change-Id: I87467600fd5e5b71bba5f8b639fca64557077058
Signed-off-by: Honghui Zhang <honghui.zhang@mediatek.com>
CR-Id: ALPS02767202
Feature: DMA