[ALPS03192728] posix_acl: Clear SGID bit when setting file permissions
When file permissions are modified via chmod(2) and the user is not in
the owning group or capable of CAP_FSETID, the setgid bit is cleared in
inode_change_ok().  Setting a POSIX ACL via setxattr(2) sets the file
permissions as well as the new ACL, but doesnt clear the setgid bit in
a similar way; this allows to bypass the check in chmod(2).  Fix that.

MTK-Commit-Id: 8cc20d11b3dccef4a4f2aee6e5eeafac8c4c91e3

Change-Id: I2aaa00047c8a9d59a2753604da2f7e2b0fcdb25b
References: CVE-2016-7097
Reviewed-by: Christoph Hellwig <hch@lst.de>
Reviewed-by: Jeff Layton <jlayton@redhat.com>
Signed-off-by: Jan Kara <jack@suse.cz>
Signed-off-by: Andreas Gruenbacher <agruenba@redhat.com>
CR-ID: ALPS03192728