[ALPS03044887] ASoC: mt6763 dctrim refactor
[Detail]
Remove Dctrim of AP side control.
New Dctrim is implemented in Pmic side.

The Mixer control of Dctrim is update
DcTrim_DC_Offset -> get/set the dc offset value
Dctrim_Control_Switch -> Control Dctrim of three state
"Not Yet", "Calibrated" show the status of Dctrim finished or not.
if set "Calibrating" to this, it will trigger the hardware to do
the dc calibration and update the dc_offset and hp_trimcode, the flag
also change to "Calibrated".

MTK-Commit-Id: 21a7ff883427325183b0249be115a298a231cc8b

Change-Id: Ic5c1749b3e5aee80378651ca30240c01bf14500f
Signed-off-by: KaiChieh Chuang <kaichieh.chuang@mediatek.com>
CR-Id: ALPS03044887
Feature: Phone Sound