[ALPS03254794] charger: return dummy temp when no VBUS
[Detail]
If USB cable is not plugged in and is not charging, return
dummy value to thermal driver for charging ICs temperature.
This will not trigger ADC measurement for charging IC and
can lower power consumption.
Only measure and return charging ICs temperature when USB
cable is plugged in.

MTK-Commit-Id: f32e705ea659056dbd9b17564b4062db780448c9

Change-Id: I867c2b21d2e698e62d666f7099f0c50c7f298d31
Signed-off-by: Elvin Lin <elvin.lin@mediatek.com>
CR-Id: ALPS03254794
Feature: Charger