[ALPS03435275] wlan: TDLS setup with keylen=0
[Detail]
In our TDLS setup flow, driver will check if the key length
is not 0. If the key is empty, driver will return and fail
the connection setup. Therefore, if two STAs are under an open
system and try the setup TDLS connection with each other,
driver will cancel the setup due to key length is 0.

[Solution]
If the key length is empty, only print some log and not
pass the key to FW; Otherwise, execute the original
procedures.

MTK-Commit-Id: a322872e06a4aaf5b1da3e0f436f623eb1f50255

Change-Id: Ia02a124aa1088b0a70bb777adeb4a72eeac3d9b3
CR-Id: ALPS03435275
Feature: Wifi Driver
Signed-off-by: mtk11918 <wayne-tw.chen@mediatek.com>