[ALPS03585694] usb: clear req_init at USB RESET interrupt
1. Driver allocate ep0 request at composite_dev_prepare()
2. Driver free ep0 request at composite_dev_cleanup()
3. Request is added to req_list at musb_g_ep0_queue()
4. Request is removed from req_list at musb_g_ep0_giveback()

This issue is the request was freed. But the request
is not removed from req_list. At the next enumeration,
the driver check there is a req in the req_list.So
want to remove from req_list. But the request was
already freed, can not be used.

The solution is to reinitialize the req_list at musb_g_reset()
when receiving MUSB_INTR_RESET.

MTK-Commit-Id: 8e03b5485bbdf812dd0fba5b25e580ae5c991d4e

Change-Id: Ic912320132d2466967ab5ad56055c171414aab0a
Signed-off-by: Justin.Hsieh <justin.hsieh@mediatek.com>
CR-Id: ALPS03585694
Feature: [Android Default] USB Tethering