[ALPS03610708] [PATCH] ANDROID: usb: gadget: f_mtp: Return error if count is negative
If the user passes in a negative file size in a int64,
this will compare to be smaller than buffer length,
and it will get truncated to form a read length that
is larger than the buffer length.

To fix, return -EINVAL if the count argument is negative,
so the loop will never happen.

MTK-Commit-Id: 349ef30dd11dbb7138476b5e2f7ce721407173e5

Bug: 37429972
Test: Test with PoC
Change-Id: I5d52e38e6fbe2c17eb8c493f9eb81df6cfd780a4
Signed-off-by: Jerry Zhang <zhangjerry@google.com>
Signed-off-by: Eddie Hung <eddie.hung@mediatek.com>
CR-Id: ALPS03610708
Feature: Others