[ALPS03740093] DISP: add condition for bypass OVL
[Detail]
When the round corner is enable, the condition of
bypass OVL only checks 0-9 layers. However, when
assert layer is enable, it is in 11th layer and it
will be missed to check by the bypass OVL condition.
Then, the path will enter RDMA mode and assert layer
will disappear.

[Solution]
Add a new condition to check whether the assert layer
is enable

MTK-Commit-Id: 75ab0f55c9e766057cda5c2dec52f8747294190e

Change-Id: I929c1f2a15e5b199e7f7d7a17e8a4be2cf55fd18
Signed-off-by: mtk14518 <freya.li@mediatek.com>
CR-Id: ALPS03740093
Feature: Display Driver (DSI Interface)