[ALPS04305265] binder: create node flag to request senders security context
To allow servers to verify client identity, allow a node
flag to be set that causes the senders security context
to be delivered with the transaction. The BR_TRANSACTION
command is extended in BR_TRANSACTION_SEC_CTX to
contain a pointer to the security context string.

MTK-Commit-Id: d9cc40ce64d1e77eb85423c12ab3f82d31f3d5ad

Change-Id: I434f6734e7491145ca17a6f9b5dd6d7d43aa7fdf
CR-Id: ALPS04305265
Feature: [Module]Binder
Signed-off-by: Todd Kjos <tkjos@google.com>
Signed-off-by: Jian-Min Liu <jian-min.liu@mediatek.com>
(cherry-pick from 67ddcd7e9561ca84971e26f77f225494d7fc44b5)