[ALPS04400710] nl80211: NL80211_ATTR_SOCKET_OWNER for CMD_CONNECT
Disconnect or deauthenticate when the owning socket is closed if this
flag is supplied to CMD_CONNECT or CMD_ASSOCIATE.  This may be used
to ensure userspace daemon doesnt leave an unmanaged connection behind.

In some situations it would be possible to account for that, to some
degree, in the daemon restart code or in the up/down scripts without
the use of this attribute.  But there will be systems where the daemon
can go away for varying periods without a warning due to local resource
management.

MTK-Commit-Id: 59a13ffbf325e3fd543b589404cf7c846612667e

Change-Id: Ibc6ffc56214b678281f08b56feb340abf923ff59
Signed-off-by: Xiang Wu <xiang.wu@mediatek.com>
Feature: Wi-Fi Certification : TGn STA mode
CR-Id: ALPS04400710