[ALPS04329470] mmc: add write protection check
When eMMC is power on write protected, some partitions
can not be written and should be set as read only.

1.Check write-protect status on BLKROSET ioctl
Check write-protect status of MMC device when performing BLKROSET ioctl
with arg 0 (for clearing read-only). If the target block device is fully
or partially write-protected, return -EACCES to block layer.
Block layer will abort the remaining BLKROSET flow when receiving
-EACESS from underlying layer and then return -EACCESS to user space
application.

With this, "adb remount" will receive error when remounted filesystem is
in write-protect region of an eMMC device.

2.Set partition as readonly according to wp status
When creating block device for eMMC logical partitions, query the
write-protect status of each partitions area in eMMC device and set
the partition as readonly if the area is fully or partially
write-protected.
With this, any raw block device write to readonly partition will get
"Opertoin not permitted" error.

The patch helps to maintain the consistency between
the device WP and the driver status.

MTK-Commit-Id: fef3ce7bf44261fe8b4aaa291a857efca8d994a4

Change-Id: I050624dac01a63ec6f638d67c89ae1738fbf69e8
Signed-off-by: Andy Teng <andy.teng@mediatek.com>
CR-Id: ALPS04329470
Feature: EMMC Partition write protection