[ALPS04756191] Sensor: Make Sure IPI_CHRE Ready before IPI Xfer
1.Using the SCP_nanohub_monitor_notify() to get SCP_READY signal,
instead of the SCP providing, to make sure the IPI_CHRE channel
is available before the IPI_CHRE xfer.

2.Replace list_add() with list_add_tail() to register sub-module
notify callback, to make sure IPI_CHRE is the first node in the
list and is executed first.

3.Add the weak function of scp_power_monitor_register() to
avoid build error when the MTK_SENSOR_SUPPORT is not set

MTK-Commit-Id: ac7bf7b7422675aeb18d3da5495a78a05bc1c8f3

Change-Id: Ibb367c3fdf0022745cf49f237a47e26e258034eb
Signed-off-by: zhichao.liu <zhichao.liu@mediatek.com>
CR-Id: ALPS04756191
Feature: Sensor Hub