[ALPS04985868] usb: gadget: core: unmap request from DMA only if previously mapped
In the SG case this is already handled since a non-zero
request->num_mapped_sgs is a clear indicator that dma_map_sg()
had been called. While it would be nice to do the same for the
singly mapped case by simply checking for non-zero request->dma,
its conceivable that 0 is a valid dma_addr_t handle. Hence add
a flag dma_mapped to struct usb_request and use this to
determine the need to call dma_unmap_single(). Otherwise, if a
request is not DMA mapped then the result of calling
usb_request_unmap_request() would safely be a no-op.

MTK-Commit-Id: a7492a41c24e0b906b5bcd9d59dcce42964a9cfc

Change-Id: Idc467944322615e4e70a2f3c6dc90ecd061fd177
Signed-off-by: Jack Pham <jackp@codeaurora.org>
Signed-off-by: Felipe Balbi <felipe.balbi@linux.intel.com>
Signed-off-by: Macpaul Lin <macpaul.lin@mediatek.com>
CR-Id: ALPS04985868
Feature: [Module]USB 2.0